test: operationalize tiers — vitest projects + per-tier scripts + move egress e2e
Encode the test tier in the filename, routed to its own vitest project, mirroring
the vigiles/unit·/integration·/e2e import paths:
- unit = src/**/*.test.ts minus the suffixed tiers (pure, no caps)
- integration = *.integration.test.ts (bwrap, no network)
- e2e = *.e2e.test.ts (real egress/model/network; self-skips where unavailable)
- runners unchanged

Move the two egress-allow tests out of run-hook.test.ts into
src/run-hook.e2e.test.ts, importing from the e2e barrel (./e2e.js) — dogfooding
the tiered import path: an e2e test pulls from vigiles/e2e, a unit test couldn't.

Scripts: test:unit / test:integration / test:e2e (vitest projects); the old
live-claude bash e2e renamed test:e2e → test:cli-e2e (docs updated).
+ coverage run all projects (e2e self-skips in CI), so the 100% gate holds
(verified: 100% stmts/funcs/lines, branches 93.8%, 756 tests).

Note: run-hook.test.ts still holds a live-net recordEgress dogfood + bwrap
confinement tests — fuller re-bucket into integration/e2e is a follow-up.
