+++ package.json
@@ -55,7 +55,10 @@
     "fmt": "prettier --write .",
     "fmt:check": "prettier --check .",
     "demo": "npm run build && bash examples/demo/run.sh",
-    "test:e2e": "bash test/e2e/run.sh",
+    "test:unit": "npm run build && vitest run --project unit",
+    "test:integration": "npm run build && vitest run --project integration",
+    "test:e2e": "npm run build && vitest run --project e2e",
+    "test:cli-e2e": "bash test/e2e/run.sh",
     "test:harness": "npm run build && node dist/cli.js test",
     "test:eval": "npm run build && node dist/cli.js eval",
     "test:vitest": "npm run build && vitest run --project runners",
